\name{cpSpecHpcMzXml}
\docType{data}
\alias{cpSpecHpcMzXml}
\title{Mass spectrum generated by Bruker Daltonics' CompassXport}
\description{
This dataset was generated by Bruker Daltonics' CompassXport and imported by
\code{\link[readMzXmlData]{readMzXmlFile}} to R. It is only needed for 
comparison between Bruker Daltonics' HPC and \code{\link[readBrukerFlexData]{.hpc}}.
}
\usage{cpSpecHpcMzXml}
\format{A list containing a mass and a intensity vector.}
\source{Examples/hpc/mzXML/hpc.mzXML}
\seealso{
    \code{\link[readBrukerFlexData]{.hpc}},
    \code{\link[readMzXmlData]{readMzXmlFile}}
}
\keyword{datasets}
