% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntbcal-functions.R
\name{.ctof2calibration}
\alias{.ctof2calibration}
\title{Cubic calibration}
\usage{
.ctof2calibration(tof, d)
}
\arguments{
\item{tof}{\code{double}, time of flight.}

\item{d}{\code{double}, calibration constants, return value of
\code{.extractV10CTOF2CalibrationConstants}.}
}
\value{
\code{double}, mz values.
}
\description{
Only basic support (not 100\% identical results) for Bruker Daltonics' cubic
calibration (encoded in "##\$NTBCal" as "V1.0CTOF2CalibrationConstants").
This is an internal function and should normally not used by the user.
}
\note{
Bruker Daltonics doesn't explain how the cubic calibration works.
All formula are results of \dQuote{trial and error}.
That is why mass calculated by \code{.ctof2calibration}
differs little from original mass.
See also \url{https://github.com/sgibb/readBrukerFlexData/issues/3}.
}
\seealso{
\url{https://github.com/sgibb/readBrukerFlexData/issues/3}
}
\author{
Alan Race, Samuel Granjeaud, Sebastian Gibb
}
\keyword{internal}
