% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{readBrukerFlexData-defunct}
\alias{mqReadBrukerFlex}
\alias{readBrukerFlexData-defunct}
\title{Removed functions in package \pkg{readBrukerFlexData}}
\description{
These functions are defunct and no longer available.
}
\details{
\describe{
 \item{mqReadBrukerFlex:}{use
   \code{\link[MALDIquantForeign]{importBrukerFlex}} instead.}
}
}
\keyword{internal}

