\name{readBrukerFlexData-internal}

\alias{.changePrecision}
\alias{.readAcquFile}
\alias{.readFidFile}
\alias{.grepAcquValue}
\alias{.extractHPCConstants}
\alias{.tof2mass}

\title{Internal readBrukerFlexData Functions}
\description{
Internal readBrukerFlexData functions.
}
\note{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written).
}
\keyword{internal}
