\encoding{utf8}
\name{readBrukerFlexData-package}
\alias{readBrukerFlexData-package}
\docType{package}
\title{The readBrukerFlexData Package}

\description{
The package reads mass spectrometry data in Bruker Daltonics' XMASS format
generated by Bruker Daltonics' mass spectrometer of *flex series. \cr
The package was developed \emph{without} any knowledge nor even support by 
Bruker Daltonics.
\cr \cr
All trademarks are owned by or licensed to Bruker Daltonics.
}

\details{
\tabular{ll}{
Package: \tab readBrukerFlexData\cr
Type: \tab Package\cr
Version: \tab 1.2.4\cr
Date: \tab 2012-02-02\cr
License: \tab GPL (version 3 or later)\cr
}    

Main functions:

\code{\link[readBrukerFlexData]{readBrukerFlexFile}}: Reads mass spectrometry data in 
Bruker Daltonics' XMASS format (fid and acqu file).

\code{\link[readBrukerFlexData]{readBrukerFlexDir}}: Reads recursively mass spectrometry data in
Bruker Daltonics' XMASS format. 

\code{\link[readBrukerFlexData]{mqReadBrukerFlex}}: Reads mass spectrometry data into MALDIquant.
}

\author{Sebastian Gibb 

Maintainer: Sebastian Gibb \email{mail@sebastiangibb.de}}
\references{
See website: \url{http://strimmerlab.org/software/maldiquant/}
}
\keyword{IO}
\seealso{
    \code{\link[readBrukerFlexData]{readBrukerFlexDir}}, 
    \code{\link[readBrukerFlexData]{readBrukerFlexFile}}
    \code{\link[readBrukerFlexData]{mqReadBrukerFlex}}
}
