\encoding{utf8}
\name{readBrukerFlexData-package}
\alias{readBrukerFlexData-package}
\docType{package}
\title{The readBrukerFlexData Package}

\description{
The package reads mass spectrometry data in Bruker Daltonics' XMASS format
generated by Bruker Daltonics' mass spectrometer of *flex series. \cr
The package was develop \emph{without} any knowledge nor even support by 
Bruker Daltonics.
\cr \cr
All trademarks are owned by or licensed to Bruker Daltonics.
}

\details{
\tabular{ll}{                                                                                                                                     
Package: \tab readBrukerFlexData\cr                                                                                                                   
Type: \tab Package\cr                                                                                                                             
Version: \tab 1.0.1\cr                                                                                                                              
Date: \tab 2011-02-16\cr                                                                                                                          
License: \tab GPL(version 3 or later)\cr                                                                                                          
}    

Main functions:

\code{\link{readBrukerFlexFile}}: Reads mass spectrometry data in 
Bruker Daltonics' XMASS format (fid and acqu file).

\code{\link{readBrukerFlexDir}}: Reads recursively mass spectrometry data in
Bruker Daltonics' XMASS format. 
}

\author{Sebastian Gibb 

Maintainer: Sebastian Gibb \email{mail@sebastiangibb.de}}
\references{
See website: \url{http://www.sebastiangibb.de/proteomics/}
}
\keyword{IO}
\seealso{
    \code{\link{readBrukerFlexDir}}, 
    \code{\link{readBrukerFlexFile}}
}
