\docType{data}
\name{cpSpecHpcMzXml}
\alias{cpSpecHpcMzXml}
\title{Mass spectrum generated by Bruker Daltonics CompassXport}
\format{A list containing a mass and an intensity vector.}
\source{
  Examples/hpc/mzXML/hpc.mzXML
}
\usage{
  cpSpecHpcMzXml
}
\description{
  This dataset was generated by Bruker Daltonics
  CompassXport and imported by
  \code{\link[readMzXmlData]{readMzXmlFile}} to R. It is
  only needed for comparison between Bruker Daltonics' HPC
  and \code{\link[readBrukerFlexData]{.hpc}}.
}
\seealso{
  \code{\link[readBrukerFlexData]{.hpc}},
  \code{\link[readMzXmlData]{readMzXmlFile}}
}
\keyword{datasets}

