% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mqReadBrukerFlex}
\alias{mqReadBrukerFlex}
\alias{mqReadBrukerFlex-deprecated}
\title{Reads mass spectrometry data into MALDIquant.}
\usage{
mqReadBrukerFlex(path, ...)
}
\arguments{
\item{path}{\code{character}, path to \emph{directory} or a single
\emph{fid} file.}

\item{\ldots}{arguments to be passed to
 \code{\link[readBrukerFlexData]{readBrukerFlexDir}} or
 \code{\link[readBrukerFlexData]{readBrukerFlexFile}}}
}
\value{
a \code{list} of \code{\link[MALDIquant]{MassSpectrum-class}} objects.
}
\description{
This function is deprecated and will be removed in the next release.
Use \code{\link[MALDIquantForeign]{importBrukerFlex}} instead. \cr
This function reads all mass spectrometry data in
Bruker Daltonics XMASS format in a specified path into
\code{\link[MALDIquant]{MALDIquant-package}}
\code{\link[MALDIquant]{MassSpectrum-class}} objects.
}
\details{
See \code{\link[readBrukerFlexData]{readBrukerFlexDir}} or
\code{\link[readBrukerFlexData]{readBrukerFlexFile}}.
}
\references{
See website: \url{http://strimmerlab.org/software/maldiquant/}
}
\seealso{
\code{\link[readBrukerFlexData]{readBrukerFlexDir}},
\code{\link[readBrukerFlexData]{readBrukerFlexFile}},
\code{\link[MALDIquant]{MALDIquant-package}},
\code{\link[MALDIquant]{MassSpectrum-class}}
}
\keyword{IO}

