% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readBrukerFlexData-defunct}
\alias{mqReadBrukerFlex}
\alias{readBrukerFlexData-defunct}
\title{Removed functions in package \pkg{readBrukerFlexData}}
\description{
These functions are defunct and no longer available.
}
\details{
\describe{
 \item{mqReadBrukerFlex:}{use
   \code{\link[MALDIquantForeign]{importBrukerFlex}} instead.}
}
}
\keyword{internal}

