% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{readBrukerFlexData-package}
\alias{readBrukerFlexData}
\alias{readBrukerFlexData-package}
\title{The readBrukerFlexData Package}
\description{
The readBrukerFlexData package reads data files acquired by MALDI-TOF MS on
Bruker Daltonics machines of the *flex series. (autoflex,
microflex, ultraflex). \cr
The package was developed \emph{without} any knowledge nor even support by
Bruker Daltonics.
\cr \cr
All trademarks are owned by or licensed to Bruker Daltonics.
}
\references{
\url{https://github.com/sgibb/readBrukerFlexData}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://strimmerlab.github.io/software/maldiquant/}
  \item \url{https://github.com/sgibb/readBrukerFlexData/}
  \item Report bugs at \url{https://github.com/sgibb/readBrukerFlexData/issues/}
}

}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{package}
