\docType{package}
\name{readBrukerFlexData-package}
\alias{readBrukerFlexData-package}
\title{The readBrukerFlexData Package}
\description{
  The package reads mass spectrometry data in Bruker
  Daltonics' XMASS format generated by Bruker Daltonics'
  mass spectrometer of *flex series (autoflex, microflex,
  ultraflex). \cr The package was developed \emph{without}
  any knowledge nor even support by Bruker Daltonics. \cr
  \cr All trademarks are owned by or licensed to Bruker
  Daltonics.
}
\details{
  \tabular{ll}{

  Package: \tab readBrukerFlexData \cr

  Version: \tab 1.4\cr

  Date: \tab 2012-09-10\cr

  License: \tab GPL (>= 3)\cr

  URL: \tab
  \url{https://github.com/sgibb/readBrukerFlexData}\cr

  }
}
\author{
  Sebastian Gibb <\email{mail@sebastiangibb.de}>
}
\references{
  \url{https://github.com/sgibb/readBrukerFlexData}
}
\keyword{package}

